backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
deepface>=0.0.93
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-multipart==0.0.9
Pillow==10.3.0
tf-keras
